在获取博客列表和搜索博客当中添加了博客总数，在获取博客列表中还添加了当前用户对这些博客的点赞收藏状态

+++ blog_main/src/main/java/cn/sticki/blog/pojo/vo/BlogListVO.java
@@ -10,10 +10,10 @@ public class BlogListVO {
 
 	List<BlogBasic> blogList;// 博客信息列表
 
-	Integer page;// 当前页
+	Long page;// 当前页
 
-	Integer pageSize;// 页大小
+	Long pageSize;// 页大小
 
-	Integer total;// 博客总数
+	Long total;// 博客总数
 
 }
